# openapi/openapi.yaml
openapi: 3.0.3
info:
  description: Self-service eSIM API. Public REST under `/api/v1/`. Staff Operations Dashboard under `/api/v1/admin/`
    (read-only). Staff internal tools under `/api/internal/`. Authenticate with JWT Bearer tokens from
    `/api/v1/auth/token/`.
  title: RoamKit API
  version: 1.0.0
servers:
- description: Staging
  url: https://api.staging.roamkit.net
- description: Production
  url: https://api.roamkit.net
...
paths:
...
  /api/v1/billing/partner-grants/:
    post:
      description: Move credits from the caller's single partner-channel team account to an attributed
        customer's personal account. The body is only customer_id, amount, and idempotency_key.
      operationId: billing_partner_grant
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PartnerGrantRequestRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartnerGrantResponse'
          description: Created grant, or the original grant on replay
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartnerGrantCode'
          description: invalid_request or invalid_amount
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartnerGrantCode'
          description: authentication_required
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartnerGrantCode'
          description: partner_access_denied or partner_grant_forbidden
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartnerGrantCode'
          description: partner_channel_disabled or customer_not_found
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartnerGrantCode'
          description: partner_context_ambiguous, idempotency_key_conflict, insufficient_funds, or customer_attribution_changed
        '429':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartnerGrantCode'
          description: rate_limited
      security:
      - bearerAuth: []
      summary: Grant partner credits to an attributed customer
      tags:
      - Billing
...
  /api/v1/orgs/partner/grants/:
    get:
      description: Stored credit grants for the caller's single partner channel.
      operationId: orgs_partner_grants
      parameters:
      - in: query
        name: order
        schema:
          enum:
          - asc
          - desc
          type: string
      - in: query
        name: page
        schema:
          type: integer
      - in: query
        name: page_size
        schema:
          type: integer
      - in: query
        name: sort
        schema:
          enum:
          - amount
          - created_at
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartnerGrantsPage'
          description: Grant page
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartnerChannelCode'
          description: invalid_page, invalid_page_size, invalid_sort, or invalid_order
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartnerChannelCode'
          description: authentication_required
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartnerChannelCode'
          description: partner_access_denied
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartnerChannelCode'
          description: partner_channel_disabled
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartnerChannelCode'
          description: partner_context_ambiguous
      security:
      - bearerAuth: []
      summary: Partner channel grant history
      tags:
      - Organizations
...
components:
  schemas:
...
    PartnerChannelCode:
      description: Partner read error body. Exactly one code, no field names.
      properties:
        code:
          type: string
      required:
      - code
      type: object
...
    PartnerGrantActor:
      properties:
        display_name:
          type: string
        email:
          type: string
        user_id:
          format: int64
          maximum: 9223372036854775807
          minimum: 1
          type: integer
      required:
      - display_name
      - email
      - user_id
      type: object
    PartnerGrantCode:
      description: Partner grant error body. Exactly one code, no field names.
      properties:
        code:
          type: string
      required:
      - code
      type: object
    PartnerGrantHistory:
      properties:
        amount:
          format: decimal
          pattern: ^-?\d{0,14}(?:\.\d{0,6})?$
          type: string
        created_at:
          format: date-time
          type: string
        customer_id:
          format: int64
          maximum: 9223372036854775807
          minimum: 1
          type: integer
        display_name:
          type: string
        email:
          nullable: true
          type: string
        grant_id:
          format: uuid
          type: string
        granted_by:
          allOf:
          - $ref: '#/components/schemas/PartnerGrantActor'
          nullable: true
      required:
      - amount
      - created_at
      - customer_id
      - display_name
      - email
      - grant_id
      - granted_by
      type: object
    PartnerGrantRequestRequest:
      description: 'Locked grant body: ``customer_id``, ``amount``, ``idempotency_key``.'
      properties:
        amount:
          minLength: 1
          type: string
        customer_id:
          format: int64
          maximum: 9223372036854775807
          minimum: 1
          type: integer
        idempotency_key:
          maxLength: 128
          minLength: 1
          type: string
      required:
      - amount
      - customer_id
      - idempotency_key
      type: object
    PartnerGrantResponse:
      properties:
        amount:
          format: decimal
          pattern: ^-?\d{0,14}(?:\.\d{0,6})?$
          type: string
        created_at:
          format: date-time
          type: string
        customer_id:
          format: int64
          maximum: 9223372036854775807
          minimum: 1
          type: integer
        grant_id:
          format: uuid
          type: string
      required:
      - amount
      - created_at
      - customer_id
      - grant_id
      type: object
    PartnerGrantsPage:
      properties:
        count:
          minimum: 0
          type: integer
        page:
          minimum: 1
          type: integer
        page_size:
          maximum: 100
          minimum: 1
          type: integer
        results:
          items:
            $ref: '#/components/schemas/PartnerGrantHistory'
          type: array
      required:
      - count
      - page
      - page_size
      - results
      type: object
...
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      scheme: bearer
      type: http